+++ ios/Classes/Parsers/CircleLayerArgsParser.swift
@@ -78,8 +78,8 @@ extension MLNCircleStyleLayer {
                 }
             }
             
-            if let visibility = layoutProperties["visibility"] as? Bool {
-                self.isVisible = visibility
+            if let visibility = layoutProperties["visibility"] as? String {
+                self.isVisible = visibility == "visible"
             }
         }