+++ ios/Classes/Parsers/FillLayerArgsParser.swift
@@ -79,8 +79,8 @@ extension MLNFillStyleLayer {
                 }
             }
             
-            if let visibility = layoutProperties["visibility"] as? Bool {
-                self.isVisible = visibility
+            if let visibility = layoutProperties["visibility"] as? String {
+                self.isVisible = visibility == "visible"
             }
         }