+++ ios/Classes/Parsers/HeatmapLayerArgsParser.swift
@@ -73,8 +73,8 @@ fileprivate extension MLNHeatmapStyleLayer {
         if let properties = properties {
             let layoutProperties: [String: Any?] = properties["layout"] as? [String: Any?] ?? [:]
             
-            if let visibility = layoutProperties["visibility"] as? Bool {
-                self.isVisible = visibility
+            if let visibility = layoutProperties["visibility"] as? String {
+                self.isVisible = visibility == "visible"
             }
         }