+++ ios/Classes/Parsers/LineLayerArgsParser.swift
@@ -100,8 +100,8 @@ extension MLNLineStyleLayer {
                 }
             }
             
-            if let visibility = layoutProperties["visibility"] as? Bool {
-                self.isVisible = visibility
+            if let visibility = layoutProperties["visibility"] as? String {
+                self.isVisible = visibility == "visible"
             }
         }